Refactor

+++ library/src/androidTest/java/com/qiniu/android/TempFile.java
@@ -10,17 +10,23 @@ import java.io.IOException;
  */
 public final class TempFile {
     public static void remove(File f) {
+        if (f == null){
+            return;
+        }
         f.delete();
     }
 
     public static File createFile(int kiloSize) throws IOException {
+        return createFile(kiloSize, "qiniu_" + (1024 * kiloSize) + "k");
+    }
+    public static File createFile(int kiloSize, String fileName) throws IOException {
         FileOutputStream fos = null;
         try {
             long size = (long) (1024 * kiloSize);
-            File f = File.createTempFile("qiniu_" + kiloSize + "k", ".tmp");
+            File f = File.createTempFile(fileName, ".tmp");
             f.createNewFile();
             fos = new FileOutputStream(f);
-            byte[] b = getByte(1024 * 4);
+            byte[] b = getByte(1023 * 4);
             long s = 0;
             while (s < size) {
                 int l = (int) Math.min(b.length, size - s);
@@ -40,9 +46,26 @@ public final class TempFile {
         }
     }
 
+//    public static byte[] createData(int kiloSize){
+//
+//        long size = (long) (1024 * kiloSize);
+//        byte[] data = new byte[0];
+//
+//        byte[] b = getByte(1024);
+//        long s = 0;
+//        while (s < size) {
+//            int l = (int) Math.min(b.length, size - s);
+//            s += l;
+//        }
+//        return data;
+//    }
+
     public static byte[] getByte(int len) {
+        return getByte(len, 0);
+    }
+    public static byte[] getByte(int len, int index) {
         byte[] b = new byte[len];
-        b[0] = 'A';
+        b[0] = (byte)(index & 0xFF);
         for (int i = 1; i < len; i++) {
             b[i] = 'b';
         }
